Reversing changes to the main

+++ src/DataTypes.java
@@ -6,7 +6,7 @@ public class DataTypes {
     public static long sum(List<Integer> numbers) {
         //testing
 
-        long s = 0;
+        int s = 0;
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
